根据CSV文件创建数据库表并插入数据

+++ src/main/java/cn/mobingi/price/config/CreateTableAndInsertDataTask.java
@@ -9,6 +9,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -28,20 +29,24 @@ public class CreateTableAndInsertDataTask {
     /**
      * <p>从价格CSV文件中获取数据库表名，字段名以及数据定时任务方法</p>
      */
-    @Scheduled(cron = "0/40 * * * * ?")
+    @Scheduled(cron = "* 28 17 * * ?")
     private void createTableAndInsertDataTask() {
         try {
-            Map<String, Object> map = ReadCsvUtils.getTableNameAndTableData();
-            String tableName = (String)map.get("tableName");
-            if (!StringUtils.isEmpty(awsMapper.existTable(tableName))) {
-                awsMapper.dropTable(map);
+            List<String> csvURL = ReadCsvUtils.getCsvHttpByIndex();
+            if (null != csvURL || csvURL.size() > 0) {
+                for (String url : csvURL) {
+                    Map<String, Object> map = ReadCsvUtils.getTableNameAndTableData();
+                    String tableName = (String)map.get("tableName");
+                    if (!StringUtils.isEmpty(awsMapper.existTable(tableName))) {
+                        awsMapper.dropTable(map);
+                    }
+                    awsMapper.createTable(map);
+                    awsMapper.insertData(map);
+                }
             }
-            awsMapper.createTable(map);
-            awsMapper.insertData(map);
         } catch (Exception e) {
             e.printStackTrace();
         }
-
     }
 
 }
